amazon-bedrock-document-comparison-poc/requirements.txt
aiohttp==3.9.2
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
awscli==1.29.61
backoff==2.2.1
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.61
botocore==1.31.61
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.13
click==8.1.7
colorama==0.4.4
coloredlogs==15.0.1
dataclasses-json==0.6.1
docutils==0.16
exceptiongroup==1.1.3
fastapi==0.109.1
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.10
GitPython==3.1.41
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
langchain==0.1.11
langsmith==0.1.17
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.26.0
onnxruntime==1.17.0
overrides==7.4.0
packaging==23.2
pandas==2.1.1
Pillow==10.2.0
pip==23.3
posthog==3.0.2
protobuf==4.24.4
pulsar-client==3.4.0
pyarrow==14.0.1
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
rsa==4.7.2
s3transfer==0.7.0
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==65.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.35.0
streamlit==1.30.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.0
torch==2.2.0
torchvision==0.17.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.36.0
typer==0.9.0
typing_extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchfiles==0.20.0
websockets==11.0.3
wheel==0.38.4
yarl==1.9.2
zipp==3.17.0
